requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
groq==1.1.1
python-dotenv==1.2.1